+++ contrib/npm_module/README.md
@@ -1,2 +1,9 @@
-Pants plugins to support the NpmModules
+Pants plugins to support Npm Modules
 ================================================================================================
+
+NPM is a package manager for javascript.
+Please see the docs here https://github.com/npm/npm
+This plugin fetches node binary and supports following 3 transpilers.
+1. Less: Less is a CSS pre-processor. http://lesscss.org/
+2. R2: A CSS LTR to RTL converter. https://www.npmjs.com/package/R2
+3. RequireJs: RequireJS is a JavaScript file and module loader. http://requirejs.org/
